requirements.txt
Flask==2.3.2
requests==2.31.0
python-pptx==0.6.21
Pillow==10.3.0
beautifulsoup4==4.12.2
lxml==4.9.3
